feat: Add 2 decimal places value for percentage DTOs and corresponding JPQL queries

+++ sud-backend/src/main/java/com/example/demo/dto/EmployeesGenderByAgeRange.java
@@ -13,7 +13,7 @@ public class EmployeesGenderByAgeRange {
     private Long male;
     private Long female;
     private Long total;
-    private Long totalPercentage;
-    private Long malePercentage;
-    private Long femalePercentage;
+    private Double totalPercentage;
+    private Double malePercentage;
+    private Double femalePercentage;
 }
